requirements.txt
absl-py==0.1.9
bleach==1.5.0
cycler==0.10.0
easydict==1.7
futures==3.1.1
html5lib==0.9999999
Markdown==2.6.11
matplotlib==2.1.2
numpy==1.14.0
pandas==0.22.0
protobuf==3.5.1
pyparsing==2.2.0
python-dateutil==2.6.1
pytz==2017.3
six==1.11.0
tensorflow-gpu==1.5.0
tensorflow-tensorboard==1.5.0
Werkzeug==0.14.1
